feat: add continuos travel rendering

+++ server/src/main/java/com/hyperlogix/server/domain/Point.java
@@ -4,6 +4,6 @@ import jakarta.persistence.Embeddable;
 
 @Embeddable
 public record Point(
-                int x,
-                int y) {
+    double x,
+    double y) {
 }
